refactor(network_bridge): split P2pConnManager god-impl into submodules
The `impl P2pConnManager` block in p2p_protoc.rs had grown to ~4,600 lines
in a single file (7,600+ LOC total), making the connection manager hard to
navigate and review. This relocates cohesive method groups into focused
`p2p_protoc/` submodules without changing any behavior.

Method groups moved (each an additional `impl P2pConnManager` block):
  - migration.rs            directed-subscribe placement / contract migration
  - broadcast.rs            outbound broadcast, interest propagation, state sync
  - dispatch.rs             priority-select result routing + per-message handlers
  - connection_lifecycle.rs teardown + handshake / new-connection setup

The parent module keeps the core event loop (build / run_event_listener /
run_event_listener_with_socket) plus all shared free functions, types,
traits and constants, so external call sites and the `pub(crate)` API are
untouched.

The move is byte-identical aside from three mechanical, non-semantic edits:
  - methods called across the module boundary are marked `pub(super)`
    (Rust visibility flows parent->child, not child->parent);
  - explicit `super::` paths in the moved code that resolved to
    `network_bridge` become `super::super::` from the deeper submodule;
  - rustfmt reflowed signatures that gained the `pub(super)` prefix.

Source-scrape guards updated for the new file layout (assertions unchanged):
  - p2p_protoc.rs: the #4300 / #4359 / #4473 pins for
    `handle_broadcast_state_change` / `broadcast_state_to_peers` now
    `include_str!("p2p_protoc/broadcast.rs")`; the #4145 `production_src()`
    helper concatenates the module root with all four submodules so the
    try_send / blocking-send scans see the whole impl.
  - op_state_manager.rs: add `p2p_protoc/broadcast.rs` to EMISSION_SITES
    (covers the moved BroadcastStateChange emission for both gossip scrapes).
  - operations/update/propagation_stats.rs: repoint the `record_broadcast`
    pin to `p2p_protoc/broadcast.rs`.
  - broadcast.rs: annotate the four pre-existing blocking `bridge.send(...)
    .await` sites with `// channel-safety: ok` + per-site reason
    (broadcast_to_peers runs in a detached `tokio::spawn` in production;
    broadcast_state_to_peers is `#[cfg(simulation_tests)]` only). These are
    not new offenders — the rule-lint diff-scope only flags them because the
    file is newly added.

Verified: cargo check (default + simulation_tests,testing), clippy, the
full p2p_protoc unit-test suite, the broadcast emission/propagation pins,
and the rule-lint blocking-send checker are all green.
